Optimizations and new features
This is a full rework of post-triangulation optimization routines to
make them less confusing and haphazard. Most generated plans are now
better, but a lot of it still depends on getting lucky in the
triangulation stage.

Fixes:

- Remove hard httplib2 dependency (Fixes #15)
- Consider cooling method used when saving cache (Fixes #13)

New features:

- Use -j fiename.js to export fields in a json format understood by IITC
Draw-Tools (Fixes #17)
- Use --no-plan-cache to not load/save cache

Other changes:

- The new solving routines are a bit slower, but they generate better
plans, so I reduced the default number of iterations from 10,000 to
5,000. We are also no longer resetting the counter after each new plan
is found, because it's better usability to know when your job finishes.
If you want to crunch more plans, just run with more iterations.

Future work:

- At this point, I should fully grok what is going on in geometry.py and
Triangle.py, because it would make more sense to use ortools and
consider cooling/travel time optimizations at triangulation stage
instead of fixing them in post and hoping that we find a lucky plan.

Signed-off-by: Konstantin Ryabitsev <konstantin@linuxfoundation.org>